requirements.txt
gradio
openai
python-dotenv
PyPDF2